--- .babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": ["module:metro-react-native-babel-preset"]
-}
