+++ .gitignore
@@ -38,6 +38,7 @@ example/android/app/src/main/gen
 
 # build
 react-native-fast-image-*.tgz
+dist/
 
 # coverage reports
 coverage
