--- .npmignore
@@ -1,28 +1,0 @@
-# Docs
-docs
-
-# Build
-android/build
-ios/build
-react-native-fast-image-*.tgz
-
-# Examples
-ReactNativeFastImageExample
-ReactNativeFastImageExampleServer
-
-# Tests
-.circleci
-__snapshots__
-*.test.js
-
-# Configs
-.babelrc
-.gitignore
-.npmignore
-prettier.config.js
-yarn.lock
-babel.config.js
-jest.config.js
-
-# IDE
-.idea/
