--- prettier.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-    semi: false,
-    singleQuote: true,
-    trailingComma: 'all',
-    tabWidth: 4,
-}
